refactor(agent-worker): restructure architecture with clear module boundaries
Major restructuring of the agent-worker package for clean responsibility
separation:

- Extract context/ to top-level module (was nested under workflow/)
- Create daemon/ module from cli/server.ts (registry, handler, daemon)
- Split cli/index.ts (1400 lines) into cli/commands/* (7 command files)
- Deduplicate model maps into core/model-maps.ts (single source of truth)
- Merge duplicated initWorkflow/initWorkflowWithMentions into one function
- Remove deprecated CLI aliases (session *, down, up, ps)
- Add ARCHITECTURE.md documenting the new module structure
