+++ packages/agent-worker/src/context/file-provider.ts
